Update README.md
The objective of this project was to hide a secret text file within a BMP image using the Least Significant Bit(LSB) encoding technique. The length of the secret text was first encoded into the image, followed by embedding the text data bit by bit into the LSB of the image pixels. During the decoding process, the length of the text is extracted first and then the text is retrieved bit by bit, ultimately receaking the hidden message.